+++ package.json
@@ -45,7 +45,7 @@
   },
   "devDependencies": {
     "babel-core": "^6.2.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.2.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.3.13",
     "babel-preset-es2015-rollup": "^1.1.1",
     "expect": "^1.13.0",
     "github": "^0.2.4",
