+++ src/helpers.js
@@ -3,7 +3,15 @@ import createDraw from './helpers/create-draw';
 import createSeriesDraw from './helpers/create-series-draw';
 import types from './helpers/types';
 
+const helpers = {
+  connect,
+  createDraw,
+  createSeriesDraw,
+  types
+};
+
 export {
+  helpers as default,
   connect,
   createDraw,
   createSeriesDraw,
