+++ src/helpers/create-draw.js
@@ -1,3 +1,3 @@
 export default function createDraw(steps) {
-
+  return () => {};
 }
