+++ src/helpers/create-series-draw.js
@@ -1,3 +1,3 @@
 export default function createSeriesDraw(steps) {
-
+  return () => {};
 }
