+++ src/utils.js
@@ -1,3 +1,8 @@
 export function extend() {
 
 }
+
+const utils = {
+  extend
+};
+export default utils;
