+++ test/components/title.spec.js
@@ -1,0 +1,20 @@
+import expect from 'expect';
+import {Title} from '../../src/components/title';
+
+describe('Title', () => {
+  var context = {};
+
+  beforeEach(() => {
+    const chart = d3.select('body').append('svg').attr('id', 'chart');
+    const layer = chart.append('g');
+
+    Object.assign(context, {
+      chart,
+      layer
+    });
+  });
+  afterEach(() => {
+    context.chart.remove();
+    context = {};
+  });
+});
