+++ test/helpers/connect.spec.js
@@ -1,0 +1,6 @@
+import expect from 'expect';
+import connect from '../../src/helpers/connect';
+
+describe('connect', () => {
+
+});
