+++ test/helpers/create-draw.spec.js
@@ -1,0 +1,6 @@
+import expect from 'expect';
+import createDraw from '../../src/helpers/create-draw';
+
+describe('createDraw', () => {
+
+});
