+++ test/helpers/create-series-draw.spec.js
@@ -1,0 +1,6 @@
+import expect from 'expect';
+import createSeriesDraw from '../../src/helpers/create-series-draw';
+
+describe('createSeriesDraw', () => {
+
+});
