+++ test/helpers/types.spec.js
@@ -1,0 +1,6 @@
+import expect from 'expect';
+import types from '../../src/helpers/types';
+
+describe('types', () => {
+
+});
