+++ test/render.spec.js
@@ -1,0 +1,6 @@
+import expect from 'expect';
+import render from '../src/render';
+
+describe('render', () => {
+
+});
