+++ test/utils.spec.js
@@ -1,0 +1,6 @@
+import expect from 'expect';
+import utils from '../src/utils';
+
+describe('utils', () => {
+
+});
